Add and remove new applications from a VM. Application info (Name and Requirements) are stored in the database.
Also updated VM creation to dynamically provide users with instance
requirements based on the instance they select in the previous window.

Ex: If they select Wordpress, it will display Wordpress, Apache, MYSQL
to them. However, if they select just Apache, it will only show apache
to them. Also displays as many VMs as there are applications. This is
for a user that would want to put each application on its own VM.